requirements.txt
matplotlib
Pillow
torch
torchvision
requests
transformers
scipy
accelerate
safetensors